# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/ProductComp/Create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new productComp",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductoComp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_ProductoCompuesto"
                        }
                    }
                }
            }
        },
        "/api/ProductComp/DeleteByID/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a productComp by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-string"
                        }
                    }
                }
            }
        },
...
        "/api/ProductComp/GetByID/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-dto_ProductoComp"
                        }
                    }
                }
            }
        },
        "/api/Sucursal/Create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sucursal"
                ],
                "summary": "Create a new sucursal",
                "parameters": [
                    {
                        "description": "Sucursal data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Sucursal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Sucursal"
                        }
                    }
                }
            }
        },
        "/api/Sucursal/DeleteByID/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sucursal"
                ],
                "summary": "Delete a sucursal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sucursal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-string"
                        }
                    }
                }
            }
        },
...
        "/api/Sucursal/GetByID/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sucursal"
                ],
                "summary": "Get a sucursal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sucursal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Sucursal"
                        }
                    }
                }
            }
        },
        "/api/Sucursal/Update/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sucursal"
                ],
                "summary": "Update a sucursal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sucursal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sucursal data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Sucursal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Sucursal"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ProductoComp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "productoComponenteID": {
                    "type": "integer"
                },
                "productoPrincipalID": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.Body-dto_ProductoComp": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ProductoComp"
                }
            }
        },
        "pkg.Body-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
...
        "pkg.Body-types_ProductoCompuesto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.ProductoCompuesto"
                }
            }
        },
        "pkg.Body-types_Sucursal": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Sucursal"
                }
            }
        },
...
        "types.Producto": {
            "type": "object",
            "properties": {
                "categoriaID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imagen": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                },
                "stock": {
                    "type": "boolean"
                }
            }
        },
        "types.ProductoCompuesto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "productoComponente": {
                    "$ref": "#/definitions/types.Producto"
                },
                "productoComponenteID": {
                    "type": "integer"
                },
                "productoPrincipal": {
                    "$ref": "#/definitions/types.Producto"
                },
                "productoPrincipalID": {
                    "type": "integer"
                }
            }
        },
        "types.Sucursal": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "direccion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                }
            }
        }
    }
}